+++ addons/sqlbox/gw/sqlbox_sql.c
@@ -29,18 +29,6 @@ struct server_type *sqlbox_init_sql(Cfg *cfg)
         return res;
     }
 #endif
-#ifdef HAVE_SDB
-    res = (struct server_type *)sqlbox_init_sdb(cfg);
-    if (res) {
-        return res;
-    }
-#endif
-#ifdef HAVE_SQLITE
-    res = (struct server_type *)sqlbox_init_sqlite(cfg);
-    if (res) {
-        return res;
-    }
-#endif
 #ifdef HAVE_SQLITE3
     res = (struct server_type *)sqlbox_init_sqlite3(cfg);
     if (res) {