+++ test/test_dbpool.c
@@ -217,66 +217,6 @@ static void oracle_client_thread(void *arg)
 }
 #endif
 
-#ifdef HAVE_SQLITE
-#include <sqlite.h>
-
-static DBConf *sqlite_create_conf(Octstr *db)
-{
-    DBConf *conf;
-    conf = gw_malloc(sizeof(DBConf));
-    conf->sqlite = gw_malloc(sizeof(SQLiteConf));
-
-    conf->sqlite->file = octstr_duplicate(db);
-
-    return conf;
-}
-
-static int callback(void *not_used, int argc, char **argv, char **col_name)
-{
-    int i;
-    
-    for (i = 0; i < argc; i++) {
-        debug("",0,"SQLite: result: %s = %s", col_name[i], argv[i]);
-    }
-
-    return 0;
-}
-
-static void sqlite_client_thread(void *arg)
-{
-    unsigned long i, succeeded, failed;
-    DBPool *pool = arg;
-    char *errmsg = 0;
-
-    succeeded = failed = 0;
-
-    info(0,"Client thread started with %ld queries to perform on pool", queries);
-
-    /* perform random queries on the pool */
-    for (i = 1; i <= queries; i++) {
-        DBPoolConn *pconn;
-        int state;
-
-        /* provide us with a connection from the pool */
-        pconn = dbpool_conn_consume(pool);
-        debug("",0,"Query %ld/%ld: sqlite conn obj at %p",
-              i, queries, (void*) pconn->conn);
-
-        state = sqlite_exec(pconn->conn, octstr_get_cstr(sql), callback, 0, &errmsg);
-        if (state != SQLITE_OK) {
-            error(0, "SQLite: %s", errmsg);
-            failed++;
-        } else {
-            succeeded++;
-        }
-
-        /* return the connection to the pool */
-        dbpool_conn_produce(pconn);
-    }
-    info(0, "This thread: %ld succeeded, %ld failed.", succeeded, failed);
-}
-#endif
-
 #ifdef HAVE_SQLITE3
 #include <sqlite3.h>
 
@@ -496,10 +436,6 @@ int main(int argc, char **argv)
         info(0, "Do tests for oracle database.");
         database_type = DBPOOL_ORACLE;
     }
-    else if (octstr_case_compare(db_type, octstr_imm("sqlite")) == 0) {
-        info(0, "Do tests for sqlite database.");
-        database_type = DBPOOL_SQLITE;
-    }
     else if (octstr_case_compare(db_type, octstr_imm("sqlite3")) == 0) {
         info(0, "Do tests for sqlite3 database.");
         database_type = DBPOOL_SQLITE3;
@@ -517,7 +453,6 @@ int main(int argc, char **argv)
         case DBPOOL_ORACLE:
             bail_out = (!user || !pass || !db) ? 1 : 0;
             break;
-        case DBPOOL_SQLITE:
         case DBPOOL_SQLITE3:
             bail_out = (!db) ? 1 : 0;
             break;
@@ -547,12 +482,6 @@ int main(int argc, char **argv)
         case DBPOOL_ORACLE:
             conf = oracle_create_conf(user, pass, db);
             client_thread = oracle_client_thread;
-            break;
-#endif
-#ifdef HAVE_SQLITE
-        case DBPOOL_SQLITE:
-            conf = sqlite_create_conf(db);
-            client_thread = sqlite_client_thread;
             break;
 #endif
 #ifdef HAVE_SQLITE3